requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-docx==1.1.2
Pillow==10.3.0
PyMuPDF==1.24.2
pywin32==306; sys_platform == 'win32'
docxcompose==1.4.0
MarkupSafe==2.1.5
requests==2.32.5
python-dotenv==1.0.1
